# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Fitnexia API
  description: |
    REST API for Fitnexia marketplace — shared by web, mobile (iOS/Android), and admin panel.
    See [API.md](./API.md) for business rules and client notes.
  version: 1.0.0
  contact:
    name: Fitnexia Engineering
servers:
  - url: https://api.fitnexia.com/v1
    description: Production
  - url: https://api.staging.fitnexia.com/v1
    description: Staging
  - url: http://localhost:3000/v1
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /memberships/invites/{code}:
    get:
      tags: [Memberships]
      security: []
      summary: Preview a membership invite by code
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
            example: A1B2C3D4
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MembershipInvitePreview'
        '400':
          description: Invite expired or no longer valid
        '404':
          description: Invite not found
  /memberships/invites/{code}/accept:
    post:
      tags: [Memberships]
      summary: Accept invite and start authorization
      description: Athlete role required. Creates membership and returns Mercado Pago authorization URL.
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptMembershipInviteResponse'
        '403':
          description: Wrong role or invite email mismatch
        '409':
          description: Already a member of this club
...
  /memberships/me/{memberId}/statement:
    get:
      tags: [Memberships]
      summary: Membership statement of account
      description: Fee, due dates, payment history, and amount due (F-42).
      parameters:
        - name: memberId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MembershipStatement'
        '404':
          description: Membership not found
  /memberships/me/{memberId}/payments/{paymentId}:
    get:
      tags: [Memberships]
      summary: Get membership payment status
      description: Athlete role required. Returns a single debt/recurring payment for the member (F-42).
      parameters:
        - name: memberId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: paymentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MembershipPayment'
        '403':
          description: Payment does not belong to this membership
        '404':
          description: Membership or payment not found
  /memberships/me/{memberId}/payments/{paymentId}/sync:
    post:
      tags: [Memberships]
      summary: Sync membership payment from Mercado Pago
      description: Athlete role required. Polls Mercado Pago for pending debt payments (F-42).
      parameters:
        - name: memberId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: paymentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  synced:
                    type: boolean
                  payment:
                    $ref: '#/components/schemas/MembershipPayment'
                  memberFeeStatus:
                    $ref: '#/components/schemas/ClubMemberFeeStatus'
                  reason:
                    type: string
        '403':
          description: Payment does not belong to this membership
        '404':
          description: Membership or payment not found
  /memberships/me/{memberId}/authorize:
    post:
      tags: [Memberships]
      summary: Start recurring debit authorization
      description: Athlete role required. Returns Mercado Pago pre-approval URL (F-41).
      parameters:
        - name: memberId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MembershipAuthorizationResponse'
        '400':
          description: Authorization not required or payments not configured
        '404':
          description: Membership not found
  /memberships/me/{memberId}/pay-debt:
    post:
      tags: [Memberships]
      summary: Pay outstanding membership debt
      description: Athlete role required. Creates a one-off checkout for overdue fees (F-42).
      parameters:
        - name: memberId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MembershipDebtPaymentResponse'
        '400':
          description: No outstanding debt or payments not configured
        '404':
          description: Membership not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    Money:
      type: object
      required: [amount, currency]
      properties:
        amount:
          type: integer
          description: Minor units (cents)
        currency:
          type: string
          example: USD
...
    MembershipBillingFrequency:
      type: string
      enum: [monthly, quarterly, annual]
    MembershipPlanType:
      type: string
      enum: [individual, family]
    ClubMemberStatus:
      type: string
      enum:
        - invited
        - pending_authorization
        - active
        - pending_payment
        - overdue
        - inactive
    ClubMemberFeeStatus:
      type: string
      enum: [up_to_date, pending, overdue, inactive]
...
    MembershipPaymentStatus:
      type: string
      enum: [pending, approved, rejected, refunded]
...
    ClubMember:
      type: object
      properties:
        id:
          type: string
          format: uuid
        institutionId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        planId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/ClubMemberStatus'
        feeStatus:
          $ref: '#/components/schemas/ClubMemberFeeStatus'
        contactName:
          type: string
        contactEmail:
          type: string
          format: email
        contactPhone:
          type: string
        joinedAt:
          type: string
          format: date-time
        leftAt:
          type: string
          format: date-time
        planName:
          type: string
        institutionName:
          type: string
        nextBillingAt:
          type: string
          format: date-time
        subscriptionStatus:
          type: string
          enum: [pending_authorization, active, past_due, cancelled]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    MembershipInvitePreview:
      type: object
      properties:
        code:
          type: string
        institutionName:
          type: string
        institutionLogo:
          type: string
          format: uri
        plan:
          type: object
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
            description:
              type: string
            price:
              $ref: '#/components/schemas/Money'
            billingFrequency:
              $ref: '#/components/schemas/MembershipBillingFrequency'
            planType:
              $ref: '#/components/schemas/MembershipPlanType'
        invitedName:
          type: string
        email:
          type: string
          format: email
        expiresAt:
          type: string
          format: date-time
    MembershipPayment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        subscriptionId:
          type: string
          format: uuid
        clubMemberId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/MembershipPaymentStatus'
        amount:
          $ref: '#/components/schemas/Money'
        periodStart:
          type: string
          format: date-time
        periodEnd:
          type: string
          format: date-time
        isManual:
          type: boolean
        checkoutUrl:
          type: string
          format: uri
        createdAt:
          type: string
          format: date-time
    MembershipStatement:
      type: object
      properties:
        member:
          $ref: '#/components/schemas/ClubMember'
        plan:
          type: object
          properties:
            name:
              type: string
            price:
              $ref: '#/components/schemas/Money'
            billingFrequency:
              $ref: '#/components/schemas/MembershipBillingFrequency'
        nextDueDate:
          type: string
          format: date-time
        amountDue:
          allOf:
            - $ref: '#/components/schemas/Money'
          nullable: true
        graceDays:
          type: integer
        payments:
          type: array
          items:
            $ref: '#/components/schemas/MembershipPayment'
    MembershipAuthorizationResponse:
      type: object
      properties:
        authorizationUrl:
          type: string
          format: uri
        preapprovalId:
          type: string
    AcceptMembershipInviteResponse:
      allOf:
        - $ref: '#/components/schemas/MembershipAuthorizationResponse'
        - type: object
          properties:
            member:
              $ref: '#/components/schemas/ClubMember'
            subscriptionId:
              type: string
              format: uuid
    MembershipDebtPaymentResponse:
      type: object
      properties:
        paymentId:
          type: string
          format: uuid
        checkoutUrl:
          type: string
          format: uri
        amount:
          $ref: '#/components/schemas/Money'
...